Create a common header for common headers and types
